+++ umecaSln/UmecaApp/Models/Sincronizacion/TabletLeaveCountryDto.cs
@@ -56,7 +56,7 @@ namespace UmecaApp
 			}
 		}
 
-		public int? webId{ get; set; }
+		public long? webId{ get; set; }
 		public int? id{ get; set; }
 		public String timeAgo{ get; set; }
 		public String reason{ get; set; }
